refactor: reorganize file structure for better maintainability
Rearrange the directories and files to create a more intuitive and
easy-to-navigate project structure.
